feat(ci-context): move to swc
Migrate from babel to swc

BREAKING CHANGE: Dropped support for Node 12
