+++ packages/ci-context/.lib.swcrc
@@ -1,0 +1,30 @@
+{
+  "jsc": {
+    "target": "es2021",
+    "parser": {
+      "syntax": "typescript",
+      "decorators": true,
+      "dynamicImport": true
+    },
+    "transform": {
+      "decoratorMetadata": true,
+      "legacyDecorator": true
+    },
+    "keepClassNames": true,
+    "externalHelpers": true,
+    "loose": true
+  },
+  "module": {
+    "type": "commonjs",
+    "strict": true,
+    "noInterop": true
+  },
+  "sourceMaps": true,
+  "exclude": [
+    "./src/**/.*.spec.ts$",
+    "./**/.*.spec.ts$",
+    "./src/**/jest-setup.ts$",
+    "./**/jest-setup.ts$",
+    "./**/.*.js$"
+  ]
+}
