+++ packages/ci-context/project.json
@@ -1,10 +1,9 @@
 {
   "root": "packages/ci-context",
   "sourceRoot": "packages/ci-context/src",
-  "projectType": "library",
   "targets": {
     "build": {
-      "executor": "@nrwl/js:tsc",
+      "executor": "@nrwl/js:swc",
       "outputs": ["{options.outputPath}"],
       "options": {
         "outputPath": "dist/packages/ci-context",
@@ -32,9 +31,11 @@
       "executor": "@jscutlery/semver:version",
       "options": {
         "baseBranch": "main",
-        "versionTagPrefix": "${target}@"
+        "preset": "conventional",
+        "tagPrefix": "${projectName}@",
+        "commitMessageFormat": "chore(${projectName}): release version ${version}"
       }
     }
   },
-  "tags": ["shared"]
+  "tags": []
 }
